% File nlme/man/BIC.Rd
% Part of the nlme package for R
% Distributed under GPL 2 or later: see nlme/LICENCE

\name{BIC}
\title{Bayesian Information Criterion}
\usage{
BIC(object, \dots)
}
\alias{BIC}
\arguments{
 \item{object}{a fitted model object.}
 \item{\dots}{optional fitted model objects.}
}
\description{
  This is the function in package \pkg{stats}, re-exported so some
  other packages which attempt to import it from \pkg{nlme} still work.
}
\keyword{models}
