% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{covind}
\alias{covind}
\title{Generate a covariance pattern matrix for pdInd}
\usage{
covind(n, ...)
}
\arguments{
\item{n}{number of random effects}

\item{...}{positions in matrix to flip from 0 to 1 or 1 to 0}
}
\description{
Generate a covariance pattern matrix for pdInd
}
