% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{disp}
\alias{disp}
\title{Display values for debugging}
\usage{
disp(x, head = deparse(substitute(x)))
}
\arguments{
\item{x}{value to display}

\item{head}{heading

Displays value and its name if \code{getOption('verbose')} is TRUE.}
}
\description{
Display values for debugging
}
