% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{f2L}
\alias{f2L}
\title{Turns a factor of a PD matrix to a left-triangular factor}
\usage{
f2L(A)
}
\arguments{
\item{A}{a factor of a positive definite matrix \eqn{V}.}
}
\value{
a right-triangular factor of the same \eqn{V}.
}
\description{
This function takes a factor of a positive-definite matrix (\eqn{V = A'A}) and
returns the equivalent left-triangular factor (\eqn{V = L'L}).
}
\seealso{
\code{f2R}
}
