% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{f2R}
\alias{f2R}
\title{Turns a factor of a PD matrix to a right-triangular factor}
\usage{
f2R(A)
}
\arguments{
\item{A}{a factor of a positive definite matrix \eqn{V}.}
}
\value{
a right-triangular factor of the same \eqn{V}.
}
\description{
This function takes a factor of a positive-definite matrix ($V = A'A$ ) and
returns the equivalent right-triangular factor ($V = R'R$).
}
\examples{
A <- matrix(rnorm(9), 3,3)
R <- f2R(A)
t(A) \%*\% A - t(R) \%*\% R
}
\seealso{
\code{f2L}
}
