% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.R
\name{getParsNlme}
\alias{getParsNlme}
\title{Calculate the parameters from the fixed and random effects}
\usage{
getParsNlme(plist, fmap, rmapRel, bmap, groups, beta, bvec, b, level, N)
}
\description{
Calculate the parameters from the fixed and random effects
}
