% File nlme/man/Nitrendipene.Rd
% Part of the nlme package for R
% Distributed under GPL 2 or later: see nlme/LICENCE

\name{Nitrendipene}
\alias{Nitrendipene}
\title{Assay of nitrendipene}
\description{
    The \code{Nitrendipene} data frame has 89 rows and 4 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
        \item{activity}{
            a numeric vector
        }
        \item{NIF}{
            a numeric vector
        }
        \item{Tissue}{
            an ordered factor with levels
            \code{2} < \code{1} < \code{3} < \code{4}
        }
        \item{log.NIF}{
            a numeric vector
        }
    }
}
\source{
  Bates, D. M. and Watts, D. G. (1988), \emph{Nonlinear Regression
    Analysis and Its Applications}, Wiley, New York.
}
%\examples{}
\keyword{datasets}
