% File nlme/man/Relaxin.Rd
% Part of the nlme package for R
% Distributed under GPL 2 or later: see nlme/LICENCE

\name{Relaxin}
\alias{Relaxin}
\title{Assay for Relaxin}
\description{
    The \code{Relaxin} data frame has 198 rows and 3 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
        \item{Run}{
            an ordered factor with levels
            \code{5} < \code{8} < \code{9} < \code{3} < \code{4} < \code{2} < \code{7} < \code{1} < \code{6}
        }
        \item{conc}{
            a numeric vector
        }
        \item{cAMP}{
            a numeric vector
        }
    }
}
\source{
    Pinheiro, J. C. and Bates, D. M. (2000), \emph{Mixed-Effects Models in S
        and S-PLUS}, Springer, New York.
}
%\examples{}
\keyword{datasets}
