% File nlme/man/Remifentanil.Rd
% Part of the nlme package for R
% Distributed under GPL 2 or later: see nlme/LICENCE

\name{Remifentanil}
\alias{Remifentanil}
\title{Pharmacokinetics of remifentanil}
\description{
    The \code{Remifentanil} data frame has 2107 rows and 12 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
        \item{ID}{
            a numeric vector
        }
        \item{Subject}{
            an ordered factor
        }
        \item{Time}{
            a numeric vector
        }
        \item{conc}{
            a numeric vector
        }
        \item{Rate}{
            a numeric vector
        }
        \item{Amt}{
            a numeric vector
        }
        \item{Age}{
            a numeric vector
        }
        \item{Sex}{
            a factor with levels
            \code{Female} 
            \code{Male} 
        }
        \item{Ht}{
            a numeric vector
        }
        \item{Wt}{
            a numeric vector
        }
        \item{BSA}{
            a numeric vector
        }
        \item{LBM}{
            a numeric vector
        }
    }
}
\source{
    Pinheiro, J. C. and Bates, D. M. (2000), \emph{Mixed-Effects Models in S
        and S-PLUS}, Springer, New York.
}
%\examples{}
\keyword{datasets}
