% File nlme/man/Wheat2.Rd
% Part of the nlme package for R
% Distributed under GPL 2 or later: see nlme/LICENCE

\name{Wheat2}
\alias{Wheat2}
\title{Wheat Yield Trials}
\description{
    The \code{Wheat2} data frame has 224 rows and 5 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
        \item{Block}{
            an ordered factor with levels
            \code{4} < \code{2} < \code{3} < \code{1}
        }
        \item{variety}{
            a factor with levels
            \code{ARAPAHOE} 
            \code{BRULE} 
            \code{BUCKSKIN} 
            \code{CENTURA} 
            \code{CENTURK78} 
            \code{CHEYENNE} 
            \code{CODY} 
            \code{COLT} 
            \code{GAGE} 
            \code{HOMESTEAD} 
            \code{KS831374} 
            \code{LANCER} 
            \code{LANCOTA} 
            \code{NE83404} 
            \code{NE83406} 
            \code{NE83407} 
            \code{NE83432} 
            \code{NE83498} 
            \code{NE83T12} 
            \code{NE84557} 
            \code{NE85556} 
            \code{NE85623} 
            \code{NE86482} 
            \code{NE86501} 
            \code{NE86503} 
            \code{NE86507} 
            \code{NE86509} 
            \code{NE86527} 
            \code{NE86582} 
            \code{NE86606} 
            \code{NE86607} 
            \code{NE86T666} 
            \code{NE87403} 
            \code{NE87408} 
            \code{NE87409} 
            \code{NE87446} 
            \code{NE87451} 
            \code{NE87457} 
            \code{NE87463} 
            \code{NE87499} 
            \code{NE87512} 
            \code{NE87513} 
            \code{NE87522} 
            \code{NE87612} 
            \code{NE87613} 
            \code{NE87615} 
            \code{NE87619} 
            \code{NE87627} 
            \code{NORKAN} 
            \code{REDLAND} 
            \code{ROUGHRIDER} 
            \code{SCOUT66} 
            \code{SIOUXLAND} 
            \code{TAM107} 
            \code{TAM200} 
            \code{VONA} 
        }
        \item{yield}{
            a numeric vector
        }
        \item{latitude}{
            a numeric vector
        }
        \item{longitude}{
            a numeric vector
        }
    }
}
\source{
    Pinheiro, J. C. and Bates, D. M. (2000), \emph{Mixed-Effects Models in S
        and S-PLUS}, Springer, New York.
}
%\examples{}
\keyword{datasets}
